[#46] refactor: 리뷰 반영
