+++ src/main/java/com/ghm/giftcardfleamarket/sale/service/SaleService.java
@@ -16,7 +16,7 @@ import org.springframework.util.CollectionUtils;
 
 import com.ghm.giftcardfleamarket.brand.mapper.BrandMapper;
 import com.ghm.giftcardfleamarket.common.BaseService;
-import com.ghm.giftcardfleamarket.common.ItemBrandPair;
+import com.ghm.giftcardfleamarket.common.model.ItemBrandPair;
 import com.ghm.giftcardfleamarket.item.domain.Item;
 import com.ghm.giftcardfleamarket.item.exception.ItemNotFoundException;
 import com.ghm.giftcardfleamarket.item.mapper.ItemMapper;
